Loader Patch

+++ plugin/include/transform.h
@@ -8,6 +8,6 @@ uint16_t modified_to_utf16(uint16_t code);
 // FUN_08884680
 uint16_t sjis_to_utf16(uint16_t sjis);
 // FUN_08884724
-int binary_search(uint16_t sjis, int low, int high);
+int binary_search(uint16_t sjis, uint16_t low, uint16_t high);
 
 void init_transform();
